+++ lang/derive/extrametas/Cargo.toml
@@ -12,4 +12,3 @@ proc-macro = true
 syn = { version = "1", features = ["full", "extra-traits", "parsing"] }
 quote = "1"
 proc-macro2 = "1"
-kaptn-attribute-meta = { path = "../../attribute/meta" }
